requirements.txt
langchain
ipykernel
python-dotenv
langchain-ollama
beautifulsoup4
langchain_community
faiss-cpu
streamlit
PyPDF2
cassio 
datasets 
tiktoken
astrapy
cassandra-driver
openai
transformers
torch
langchain_openai